+++ spine-cocos2dx/example/Classes/RaptorExample.cpp
@@ -55,9 +55,10 @@ bool RaptorExample::init () {
 	
 	EventListenerTouchOneByOne* listener = EventListenerTouchOneByOne::create();
 	listener->onTouchBegan = [this] (Touch* touch, Event* event) -> bool {
-		if (!skeletonNode->getDebugBonesEnabled())
+		if (!skeletonNode->getDebugBonesEnabled()) {
 			skeletonNode->setDebugBonesEnabled(true);
-		else if (skeletonNode->getTimeScale() == 1)
+			skeletonNode->setDebugMeshesEnabled(true);
+		} else if (skeletonNode->getTimeScale() == 1)
 			skeletonNode->setTimeScale(0.3f);
 		else
 			Director::getInstance()->replaceScene(TankExample::scene());